fixed ejb

+++ wizard-ejb/src/main/java/com/engagepoint/labs/wizard/ruleExecutors/RuleExecutorAbstract.java
@@ -69,7 +69,7 @@ public abstract class RuleExecutorAbstract {
     }
 
     private boolean compareFile(Value parentQuestionAnswer, String stringToCompareWith) {
-        return ((parentQuestionAnswer.getValue()) != 0) && (stringToCompareWith.equals("true"));
+        return ((Integer)(parentQuestionAnswer.getValue()) != 0) && (stringToCompareWith.equals("true"));
     }
 
     private boolean compareList(Value parentQuestionAnswer, String[] stringArrayToCompareWith) {
